+++ backend/dataall/base/cdkproxy/stacks/__init__.py
@@ -1,7 +1,9 @@
 from .manager import stack, instanciate_stack, StackManager
+from dataall.core.environment.cdk.environment_stack import EnvironmentSetup
 
 __all__ = [
     'StackManager',
     'stack',
     'instanciate_stack',
+    "EnvironmentSetup"
 ]
